move audio converter helper
